requirements.txt
pandas
seaborn
numpy
matplotlib.pyplot
scikit-learn